Moved static vars around, commented out broken animator, many other changes
