--- src/Config.py
@@ -1,15 +1,0 @@
-import os
-
-#relative file path
-PATH = os.path.dirname(os.path.realpath(__file__))
-
-#animator
-MSPEED = 9
-KICKMOVE = 0
-
-#Level
-
-SIZE = (900, 650)
-CLOCK_SPEED = 15
-print(PATH)
-
